Fix issue with WD_LOG not working as intended (#405)
* Update logger.py

I believe your intent with `WD_LOG` is to allow users to control what level log messages are written as. As such, using `logging.setLevel()` does not do this (it only will set the level at which messages are culled). This change should fix what you intend to happen appropriately.

* Update logger.py

Co-authored-by: Sergey Pirogov <serhii.pirohov@gmail.com>